перенес функции в подходящие модули
