+++ src/components/menu.js
@@ -1,5 +1,26 @@
-const createFilmMenu = function (filmsToFilter) {
-  const {watchListCount, historyCount, favoritesCount} = filmsToFilter;
+const getFilmsToFilters = function (filmList) {
+  const filters = {
+    watchListCount: 0,
+    historyCount: 0,
+    favoritesCount: 0
+  };
+
+  filmList.forEach((element) => {
+    if (element.isInWachList) {
+      filters[`watchListCount`] += 1;
+    }
+    if (element.isInHistory) {
+      filters[`historyCount`] += 1;
+    }
+    if (element.isInFavorites) {
+      filters[`favoritesCount`] += 1;
+    }
+  });
+  return filters;
+};
+
+const createFilmMenu = function (filmList) {
+  const {watchListCount, historyCount, favoritesCount} = getFilmsToFilters(filmList);
   return (
     `<nav class="main-navigation">
       <a href="#all" class="main-navigation__item main-navigation__item--active">All movies</a>
