+++ src/components/rank.js
@@ -1,4 +1,22 @@
-const createRank = function (rank) {
+import {MinWatchedFilmsForRank} from '../const.js';
+
+
+const getRank = function (filmCount) {
+  const {NOVICE, FAN, MOVIE_BUFF} = MinWatchedFilmsForRank;
+  if (filmCount < NOVICE) {
+    return ``;
+  } else if (filmCount >= NOVICE && filmCount < FAN) {
+    return `novice`;
+  } else if (filmCount >= FAN && filmCount < MOVIE_BUFF) {
+    return `fan`;
+  } else {
+    return `movie buff`;
+  }
+};
+
+
+const createRank = function (filmCount) {
+  const rank = getRank(filmCount);
   return (
     `<section class="header__profile profile">
       <p class="profile__rating">${rank}</p>
